feat: Add WebSocket handler dependency and modify class
- Added spring-websocket dependency (build.gradle / pom.xml)
- Extended TextWebSocketHandler class and updated methods
- Improved WebSocketSession URI handling

+++ k_rstudy/src/main/kotlin/com/krstudy/kapi/domain/messages/dto/UnreadCountResponse.kt
@@ -1,5 +1,3 @@
 package com.krstudy.kapi.domain.messages.dto
 
-data class UnreadCountResponse(
-    val count: Int
-)
+data class UnreadCountResponse(val count: Long)
